6-size.c

+++ 0x00-hello_world/6-size.c
@@ -1,6 +1,7 @@
 #include <stdio.h>
 /**
  * main - Entry point
+ *
  * Return: 0(Success)
  */
 int main(void)
@@ -11,11 +12,10 @@ int main(void)
 	long long int d;
 	float g;
 
-printf("Size of a char: %d bytes(s)\n", (unsigned long)
-sizeof(a));
-printf("Size of an int: %d byte(s)\n", (unsigned long)sizeof(b));
-printf("Size of a long int: %d byte(s)\n", (unsigned long)sizeof(c));
-printf("Size of a long long int: %d byte(s)\n", (unsigned long)sizeof(d));
-printf("Size of a float: %d byte(s)\n", (unsigned long)sizeof(g));
+printf("Size of a char: %d bytes(s)\n",sizeof(a));
+printf("Size of an int: %d byte(s)\n",sizeof(b));
+printf("Size of a long int: %d byte(s)\n",sizeof(c));
+printf("Size of a long long int: %d byte(s)\n",sizeof(d));
+printf("Size of a float: %d byte(s)\n",sizeof(g));
 return (0);
 }
